+++ types/protyle.d.ts
@@ -1,7 +1,76 @@
-import {App, ILuteNode, IObject, IOperation, IPosition, IProtyleOption} from "./../siyuan";
+import {
+    App,
+    ILuteNode,
+    IObject,
+    IOperation,
+    IPosition,
+    TTurnInto,
+    TTurnIntoOne,
+    TTurnIntoOneSub
+} from "./../siyuan";
+
+type TProtyleAction = "cb-get-append" | // 向下滚动加载
+    "cb-get-before" | // 向上滚动加载
+    "cb-get-unchangeid" | // 上下滚动，定位时不修改 blockid
+    "cb-get-hl" | // 高亮
+    "cb-get-focus" | // 光标定位
+    "cb-get-focusfirst" | // 动态定位到第一个块
+    "cb-get-setid" | // 重置 blockid
+    "cb-get-all" | // 获取所有块
+    "cb-get-backlink" | // 悬浮窗为传递型需展示上下文
+    "cb-get-unundo" | // 不需要记录历史
+    "cb-get-scroll" | // 滚动到指定位置
+    "cb-get-context" | // 包含上下文
+    "cb-get-html" | // 直接渲染，不需要再 /api/block/getDocInfo，否则搜索表格无法定位
+    "cb-get-history" // 历史渲染
+
+interface IToolbarItem {
+    /** 唯一标示 */
+    name: string;
+    /** 提示 */
+    tip?: string;
+    /** svg 图标 */
+    icon: string;
+    /** 快捷键 */
+    hotkey?: string;
+    /** 提示位置 */
+    tipPosition: string;
+    click?(protyle: Protyle): void;
+}
+
+export interface IProtyleOption {
+    action?: TProtyleAction[];
+    mode?: "preview" | "wysiwyg";
+    toolbar?: Array<string | IToolbarItem>;
+    blockId?: string;
+    key?: string;
+    scrollAttr?: {
+        rootId: string,
+        startId: string,
+        endId: string,
+        scrollTop: number,
+        focusId?: string,
+        focusStart?: number,
+        focusEnd?: number,
+        zoomInId?: string,
+    };
+    defId?: string;
+    render?: {
+        background?: boolean,
+        title?: boolean,
+        gutter?: boolean,
+        scroll?: boolean,
+        breadcrumb?: boolean,
+        breadcrumbDocName?: boolean,
+    };
+    typewriterMode?: boolean;
+
+    after?(protyle: Protyle): void;
+}
 
 // REF: https://github.com/siyuan-note/siyuan/blob/dev/app/src/types/protyle.d.ts
 export interface IProtyle {
+    getInstance: () => Protyle;
     app: App;
     transactionTime: number;
     id: string;
@@ -45,7 +114,7 @@ export interface IProtyle {
     contentElement?: HTMLElement;
     options: any;
     lute?: Lute;
-    toolbar?: any;
+    toolbar?: Toolbar;
     preview?: any;
     hint?: any;
     upload?: any;
@@ -74,6 +143,33 @@ export class Protyle {
     insert(html: string, isBlock?: boolean, useProtyleRange?: boolean): void
 
     transaction(doOperations: IOperation[], undoOperations?: IOperation[]): void;
+
+    /**
+     * 多个块转换为一个块
+     * @param {TTurnIntoOneSub} [subType] type 为 "BlocksMergeSuperBlock" 时必传
+     */
+    public turnIntoOneTransaction(selectsElement: Element[], type: TTurnIntoOne, subType?: TTurnIntoOneSub): void
+
+    /**
+     * 多个块转换
+     * @param {Element} [nodeElement] 优先使用包含 protyle-wysiwyg--select 的块，否则使用 nodeElement 单块
+     * @param type
+     * @param {number} [subType] type 为 "Blocks2Hs" 时必传
+     */
+    public turnIntoTransaction(nodeElement: Element, type: TTurnInto, subType?: number): void
+
+    public updateTransaction(id: string, newHTML: string, html: string): void
+
+    public updateBatchTransaction(nodeElements: Element[], cb: (e: HTMLElement) => void): void
+
+    public getRange(element: Element): Range
+
+    public hasClosestBlock(element: Node): false | HTMLElement
+
+    /**
+     * @param {boolean} [toStart=true]
+     */
+    public focusBlock(element: Element, toStart?: boolean): false | Range
 }
 
 export class Toolbar {